fromSb3: don't re-run getBlockScript redundantly
getBlockScript returns a function which is useful for getting any
top-level script, and getBlockScript only operates on the total
list of the target's blocks, which doesn't change dependant on the
script you are trying to access. So use it like the currying
function it presents itself as and don't re-run getBlockScript
for every script in the target.