update compilesb3 snapshot test
itemOf() returns anything - just the value at that place in the
list, not casted by default - so new index rules say it should be
treated as a zero-cast number.

Of course this can cause behavior problems as indicated in #153,
but in general casting to a number before subtracting is the
correct behavior. NaN always becomes -1 when an index is desired.