Upgrade most dev tooling
Bumps eslint, typescript-eslint, and prettier. eslint now has a new
config format, which I migrated over to.

Also fixes or ignores all non-Prettier lints. Behavior should not
change.

Prettier v3 formats code asynchronously, so we can't use it with
toLeopard. Instead, we use v3 for the codebase and v2 for the API.